+++ classes/event/file_similarity_check_completed.php
@@ -14,21 +14,19 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 /**
- * unicheck_event_assessable_submited.class.php
+ * file_similarity_check_completed.php
  *
  * @package     plagiarism_unicheck
  * @subpackage  plagiarism
  * @author      Aleksandr Kostylev <a.kostylev@p1k.co.uk>
- * @copyright   UKU Group, LTD, https://www.unicheck.com
+ * @copyright   2018 UKU Group, LTD, https://www.unicheck.com
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace plagiarism_unicheck\classes\event;
+namespace plagiarism_unicheck\event;
 
-use core\event\base;
 use plagiarism_unicheck;
-use plagiarism_unicheck\classes\unicheck_adhoc;
-use plagiarism_unicheck\classes\unicheck_core;
+use plagiarism_unicheck\classes\event\abstract_file_event;
 
 if (!defined('MOODLE_INTERNAL')) {
     die('Direct access to this script is forbidden.');
@@ -37,39 +35,43 @@ if (!defined('MOODLE_INTERNAL')) {
 require_once(dirname(__FILE__) . '/../../locallib.php');
 
 /**
- * Class unicheck_event_file_submited
+ * Class file_similarity_check_completed
  *
  * @package     plagiarism_unicheck
  * @subpackage  plagiarism
+ *
  * @author      Aleksandr Kostylev <a.kostylev@p1k.co.uk>
  * @copyright   UKU Group, LTD, https://www.unicheck.com
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ * @since       Moodle 3.3
  */
-class unicheck_event_workshop_switched extends unicheck_abstract_event {
+class file_similarity_check_completed extends abstract_file_event {
     /**
-     * handle_event
+     * Init method.
      *
-     * @param unicheck_core $core
-     * @param base          $event
+     * @return void
      */
-    public function handle_event(unicheck_core $core, base $event) {
+    protected function init() {
+        $this->data['crud'] = 'u';
+        $this->data['edulevel'] = self::LEVEL_OTHER;
+        $this->data['objecttable'] = UNICHECK_FILES_TABLE;
+    }
 
-        if (!empty($event->other['workshopphase'])
-            && $event->other['workshopphase'] == UNICHECK_WORKSHOP_ASSESSMENT_PHASE
-        ) { // Assessment phase.
-            $ufiles = plagiarism_unicheck::get_area_files($event->contextid, UNICHECK_WORKSHOP_FILES_AREA);
-            $assignfiles = get_file_storage()->get_area_files($event->contextid,
-                'mod_workshop', 'submission_attachment', false, null, false
-            );
+    /**
+     * Return the event name.
+     *
+     * @return string
+     */
+    public static function get_name() {
+        return plagiarism_unicheck::trans('event:file_similarity_check_completed');
+    }
 
-            $files = array_merge($ufiles, $assignfiles);
-
-            if (!empty($files)) {
-                foreach ($files as $file) {
-                    $core->userid = $file->get_userid();
-                    unicheck_adhoc::upload($file, $core);
-                }
-            }
-        }
+    /**
+     * Returns description of what happened.
+     *
+     * @return string
+     */
+    public function get_description() {
+        return "User file '{$this->other['fileid']}' similarity check completed in course module '{$this->contextinstanceid}'";
     }
 }